fix data tables header
