+++ src/main/java/io/jenkins/plugins/reporter/ReportViewModel.java
@@ -19,6 +19,7 @@ import io.jenkins.plugins.reporter.model.Report;
 import org.apache.commons.collections.ListUtils;
 import org.kohsuke.stapler.bind.JavaScriptMethod;
 
+import javax.swing.text.TableView;
 import java.util.*;
 import java.util.stream.Collectors;
 
@@ -138,128 +139,5 @@ public class ReportViewModel extends DefaultAsyncTableContentProvider implements
     @Override
     public TableModel getTableModel(String id) {
         return new ReportTableModel(id);
-    }
-
-    /**
-     * UI table model for the report table.
-     */
-    static class ReportTableModel extends TableModel {
-        
-        private final String id;
-
-        ReportTableModel(final String id) {
-            super();
-
-            this.id = id;
-        }
-
-        @Override
-        public String getId() {
-            return id;
-        }
-
-        @Override
-        public TableConfiguration getTableConfiguration() {
-            TableConfiguration tableConfiguration = new TableConfiguration();
-            tableConfiguration.select(SelectStyle.SINGLE);
-            return tableConfiguration;
-        }
-
-        @Override
-        public List<TableColumn> getColumns() {
-            List<TableColumn> columns = new ArrayList<>();
-            
-            // this column is hidden, but used to access the file hash from the frontend
-            TableColumn fileHashColumn = new TableColumn("Hash", "fileHash");
-            fileHashColumn.setHeaderClass(TableColumn.ColumnCss.HIDDEN);
-            // fileHashColumn.setWidth(0);
-            columns.add(fileHashColumn);
-
-            TableColumn packageColumn = new TableColumn("Package", "packageName");
-            // packageColumn.setWidth(2);
-            columns.add(packageColumn);
-
-            TableColumn fileColumn = new TableColumn("File", "fileName");
-            // fileColumn.setWidth(2);
-            columns.add(fileColumn);
-
-            TableColumn lineColumn = new TableColumn("Line", "lineCoverage", "number");
-            // lineColumn.setWidth(2);
-            columns.add(lineColumn);
-
-            TableColumn lineColumnDelta = new TableColumn("Line Δ", "lineCoverageDelta", "number");
-            // lineColumnDelta.setWidth(1);
-            columns.add(lineColumnDelta);
-
-            TableColumn branchColumn = new TableColumn("Branch", "branchCoverage", "number");
-            // branchColumn.setWidth(2);
-            columns.add(branchColumn);
-
-            TableColumn branchColumnDelta = new TableColumn("Branch Δ", "branchCoverageDelta", "number");
-            // branchColumnDelta.setWidth(1);
-            columns.add(branchColumnDelta);
-
-            TableColumn loc = new TableColumn("LOC", "loc", "number");
-            // loc.setWidth(1);
-            columns.add(loc);
-
-            return columns;
-        }
-
-        @Override
-        public List<Object> getRows() {
-            List<Object> rows = new ArrayList<>();
-            rows.add(new Row(id));
-            return rows;
-        }
-
-        /**
-         * UI row model for the coverage details table.
-         */
-        private static class Row {
-
-            String id;
-            
-            Row(String id) {
-                this.id = id;
-            }
-
-            public String getFileHash() {
-                return String.valueOf(id.hashCode());
-            }
-
-            public String getFileName() {
-                return id;
-            }
-
-            public String getPackageName() {
-                return id;
-            }
-
-            public DetailedColumnDefinition getLineCoverage() {
-                return new DetailedColumnDefinition("lineCov", "lineCov");
-            }
-
-            public DetailedColumnDefinition getBranchCoverage() {
-                return new DetailedColumnDefinition("branchCov", "branchCov");
-            }
-
-            public DetailedColumnDefinition getLineCoverageDelta() {
-                return new DetailedColumnDefinition("deltCov", "deltCov");
-            }
-
-            public DetailedColumnDefinition getBranchCoverageDelta() {
-                return new DetailedColumnDefinition("branchDeltCov", "branchDeltCov");
-            }
-
-            public DetailedColumnDefinition getLoc() {
-                return new DetailedColumnDefinition("loc", "loc");
-            }
-
-           
-        }
-
-
-
     }
 }
