+++ src/main/resources/io/jenkins/plugins/reporter/ReportViewModel/index.jelly
@@ -3,11 +3,7 @@
  
   <st:header name="Content-Type" value="text/html;charset=UTF-8"/>
 
-  <bs:page it="${it}" class="fluid-container d-flex flex-column h-100">
-  
-      <st:adjunct includes="io.jenkins.plugins.jquery3"/>
-      <st:adjunct includes="io.jenkins.plugins.echarts"/>
-      <st:adjunct includes="io.jenkins.plugins.data-tables-select"/>
+  <bs:page it="${it}">
       
       <link rel="stylesheet" href="${resURL}/plugin/data-reporting/css/custom-style.css"/>
       
@@ -65,8 +61,8 @@
                                   
                                   <div class="row py-3">
                                         
-                                        <div class="col-12">
-                                        
+                                        <div class="col">
+                                            
                                             <dt:table model="${it.getTableModel(item.id)}"/>
                                             
                                         </div>
